requirements.txt
streamlit
google-cloud-storage
google-cloud-aiplatform
pydub
google-cloud-speech
python-dotenv
